requirements.txt
###### Requirements without Version Specifiers ######`
discord.py
requests
python-dotenv
Logger
rich